rename SetAutoCommitExtractor
