+++ sharding-core/src/main/java/io/shardingsphere/core/parsing/antlr/extractor/impl/PredicateExtractor.java
@@ -190,7 +190,7 @@ public final class PredicateExtractor implements OptionalSQLSegmentExtractor {
             Integer index = questionNodeIndexMap.get(expressionNode.get());
             commonExpressionSegment.setIndex(index);
         } else {
-            Optional<ParserRuleContext> bitExprNode = ExtractorUtils.findFirstChildNode(valueNode, RuleName.BITEXPR);
+            Optional<ParserRuleContext> bitExprNode = ExtractorUtils.findFirstChildNode(valueNode, RuleName.BIT_EXPR);
             expressionNode = ExtractorUtils.findFirstChildNode(valueNode, RuleName.NUMBER);
             if (expressionNode.isPresent() && (!bitExprNode.isPresent() || 1 == bitExprNode.get().getChildCount())) {
                 commonExpressionSegment.setValue(NumberUtil.getExactlyNumber(expressionNode.get().getText(), 10));
